requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pillow==10.1.0
numpy<2
ultralytics==8.0.196
python-multipart==0.0.6
